+++ src/picounits/core/quantities/factory.py
@@ -11,7 +11,7 @@ from typing import Any, Callable
 from numpy import ndarray, integer, floating, complexfloating
 from picounits.core.quantities.packet import Packet
 
-from picounits.lazy_imports import lazy_import
+from picounits.utilities.lazy_imports import lazy_import
 
 
 class Factory: