+++ src/picounits/core/quantities/scalars/types/real.py
@@ -17,7 +17,7 @@ from picounits.core.scales import PrefixScale
 from picounits.core.quantities.packet import Packet
 from picounits.core.quantities.scalars.scalar import ScalarPacket
 
-from picounits.lazy_imports import import_factory
+from picounits.utilities.lazy_imports import import_factory
 
 # Import transcendental logic functions
 from picounits.core.quantities.scalars.methods import transcendental as tlops