+++ src/picounits/core/quantities/vectors/types/array.py
@@ -22,7 +22,7 @@ from picounits.core.scales import PrefixScale
 from picounits.core.quantities.packet import Packet
 from picounits.core.quantities.vectors.vector import VectorPacket
 
-from picounits.lazy_imports import import_factory
+from picounits.utilities.lazy_imports import import_factory
 from picounits.configuration.management import get_significant_figures
 
 @dataclass(slots=True, repr=False, unsafe_hash=True)